pyproject.toml
[tool.poetry]
name = "lms"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.7"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
